Add setter methods for Sentry client
