+++ sentry.go
@@ -331,21 +331,6 @@ func (hook *SentryHook) Levels() []logrus.Level {
 	return hook.levels
 }
 
-// SetRelease sets release tag.
-func (hook *SentryHook) SetRelease(release string) {
-	hook.client.SetRelease(release)
-}
-
-// SetEnvironment sets environment tag.
-func (hook *SentryHook) SetEnvironment(environment string) {
-	hook.client.SetEnvironment(environment)
-}
-
-// SetServerName sets server_name tag.
-func (hook *SentryHook) SetServerName(serverName string) {
-	hook.serverName = serverName
-}
-
 // AddIgnore adds field name to ignore.
 func (hook *SentryHook) AddIgnore(name string) {
 	hook.ignoreFields[name] = struct{}{}
